Restructure code into modules.
